pyproject.toml
[project]
name = "defi-vuln-benchmark"
version = "0.1.0"
description = "Benchmark for evaluating AI agents on DeFi smart contract vulnerability identification"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2",
    "langchain-openai>=1.0",
    "langgraph>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = []
